\name{Reference}
\alias{Reference}
\title{The R5 Reference class}
\description{
  Pass an R object to ref. The return is an object that
  behaves like the original object but exhibits reference
  semantics. This class is useful when dealing with large
  objects and a user want to have control over when the
  object is copied.
}

