\name{solve}
\alias{solve}
\alias{solve,ANY,Reference-method}
\alias{solve,Reference,ANY-method}
\title{Solve a system of linear equations when one of the matrices is
maintained by a Reference class instance.}
\description{
  Solve a system of linear equations when one of the
  matrices is maintained by a Reference class instance.
}

